+++ CHANGELOG.md
@@ -28,6 +28,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 *   (empty)
 
+## 1.1.1 (2018-03-06)
+
+### Added
+
+*   Added support for resolving Symfony parameters using environment placeholders (Symfony 3.2+)
+*   Added rewriting Doctrine configuration key "url" (Symfony 4.0+)
+*   Added Composer scripts for PHPUnit
+*   Added Composer scripts for PHP CS Fixer
+
 ## 1.1.0 (2018-01-20)
 
 ### Added